# services/fire-viewer-backend/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fire-Viewer API",
    "description": "Transactional incident registry with idempotent ingestion, conservative spatial matching, immutable audit events and viewer manifests.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/admin/incidents": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Incidents",
        "description": "Private operational inventory keyed by the stable fire_id, never by zone_id.",
        "operationId": "list_incidents_api_v1_admin_incidents_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminIncidentListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/work-queue": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Work Queue",
        "operationId": "get_work_queue_api_v1_admin_work_queue_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminWorkQueueResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v2/admin/dashboard": {
      "get": {
        "tags": [
          "admin-v2"
        ],
        "summary": "Dashboard",
        "operationId": "dashboard_api_v2_admin_dashboard_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminDashboardResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v2/admin/work-queue": {
      "get": {
        "tags": [
          "admin-v2"
        ],
        "summary": "Work Queue",
        "operationId": "work_queue_api_v2_admin_work_queue_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminWorkQueueResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v2/admin/incidents": {
      "get": {
        "tags": [
          "admin-v2"
        ],
        "summary": "Incidents",
        "operationId": "incidents_api_v2_admin_incidents_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminIncidentListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v2/admin/operational-map": {
      "get": {
        "tags": [
          "admin-v2"
        ],
        "summary": "Operational Map",
        "description": "National internal map; each marker remains a stable incident-centred fire_id.",
        "operationId": "operational_map_api_v2_admin_operational_map_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOperationalMapResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v2/admin/incidents/{fire_id}/representations": {
      "post": {
        "tags": [
          "admin-v2"
        ],
        "summary": "Attach Representations",
        "operationId": "attach_representations_api_v2_admin_incidents__fire_id__representations_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "fire_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^FR-[0-9A-Z]{2,3}-[0-9]{5}$",
              "title": "Fire Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminIncidentRepresentationAttachRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminIncidentRepresentationAttachResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/admin/packages/{package_id}/files/{file_id}": {
      "get": {
        "tags": [
          "admin-v2"
        ],
        "summary": "Package File",
        "operationId": "package_file_api_v2_admin_packages__package_id__files__file_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "package_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 3,
              "maxLength": 96,
              "title": "Package Id"
            }
          },
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "File Id"
            }
          },
          {
            "name": "If-None-Match",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-None-Match"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/admin/assets/{asset_id}": {
      "get": {
        "tags": [
          "admin-v2"
        ],
        "summary": "Model Asset",
        "operationId": "model_asset_api_v2_admin_assets__asset_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 3,
              "maxLength": 64,
              "title": "Asset Id"
            }
          },
          {
            "name": "If-None-Match",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-None-Match"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AdminDashboardPriorityItem": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "observation",
              "report",
              "incident",
              "job",
              "model_package"
            ],
            "title": "Kind"
          },
          "priority": {
            "type": "string",
            "enum": [
              "critical",
              "high",
              "medium"
            ],
            "title": "Priority"
          },
          "target_id": {
            "type": "string",
            "title": "Target Id"
          },
          "fire_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fire Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind",
          "priority",
          "target_id",
          "title",
          "detail",
          "created_at"
        ],
        "title": "AdminDashboardPriorityItem"
      },
      "AdminDashboardQueueSummary": {
        "properties": {
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "critical": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Critical"
          },
          "high": {
            "type": "integer",
            "minimum": 0.0,
            "title": "High"
          },
          "medium": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Medium"
          },
          "observations_pending": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Observations Pending"
          },
          "reports_pending": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Reports Pending"
          },
          "incidents_requiring_review": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Incidents Requiring Review"
          },
          "jobs_quarantined": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Jobs Quarantined"
          },
          "models_to_review": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Models To Review"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "total",
          "critical",
          "high",
          "medium",
          "observations_pending",
          "reports_pending",
          "incidents_requiring_review",
          "jobs_quarantined",
          "models_to_review"
        ],
        "title": "AdminDashboardQueueSummary"
      },
...
      "AdminDashboardResponse": {
        "properties": {
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "queue": {
            "$ref": "#/components/schemas/AdminDashboardQueueSummary"
          },
          "priorities": {
            "items": {
              "$ref": "#/components/schemas/AdminDashboardPriorityItem"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Priorities"
          },
          "watchlist": {
            "items": {
              "$ref": "#/components/schemas/AdminDashboardWatchIncident"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Watchlist"
          },
          "recent_publications": {
            "items": {
              "$ref": "#/components/schemas/AdminDashboardRecentPublication"
            },
            "type": "array",
            "maxItems": 10,
            "title": "Recent Publications"
          },
          "map_summary": {
            "$ref": "#/components/schemas/AdminOperationalMapSummary"
          },
          "system": {
            "$ref": "#/components/schemas/AdminSystemStatus"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "generated_at",
          "queue",
          "map_summary",
          "system"
        ],
        "title": "AdminDashboardResponse"
      },
      "AdminDashboardWatchIncident": {
        "properties": {
          "fire_id": {
            "type": "string",
            "title": "Fire Id"
          },
          "canonical_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Name"
          },
          "status": {
            "$ref": "#/components/schemas/IncidentStatus"
          },
          "verification_state": {
            "$ref": "#/components/schemas/VerificationState"
          },
          "last_observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Observed At"
          },
          "review_required": {
            "type": "boolean",
            "title": "Review Required"
          },
          "pending_observation_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Pending Observation Count"
          },
          "model_update_available": {
            "type": "boolean",
            "title": "Model Update Available"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "fire_id",
          "status",
          "verification_state",
          "last_observed_at",
          "review_required",
          "pending_observation_count",
          "model_update_available"
        ],
        "title": "AdminDashboardWatchIncident"
      },
...
      "AdminIncidentListResponse": {
        "properties": {
          "incidents": {
            "items": {
              "$ref": "#/components/schemas/AdminIncidentSummary"
            },
            "type": "array",
            "maxItems": 200,
            "title": "Incidents"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AdminIncidentListResponse"
      },
...
      "AdminIncidentRepresentationAttachRequest": {
        "properties": {
          "package_id": {
            "type": "string",
            "maxLength": 96,
            "minLength": 3,
            "title": "Package Id"
          },
          "expected_incident_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Incident Version"
          },
          "primary_profile": {
            "type": "string",
            "enum": [
              "close",
              "local",
              "extended",
              "mobile",
              "desktop"
            ],
            "title": "Primary Profile",
            "default": "local"
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 10,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "package_id",
          "expected_incident_version",
          "reason"
        ],
        "title": "AdminIncidentRepresentationAttachRequest"
      },
      "AdminIncidentRepresentationAttachResponse": {
        "properties": {
          "fire_id": {
            "type": "string",
            "title": "Fire Id"
          },
          "episode_id": {
            "type": "string",
            "title": "Episode Id"
          },
          "package_id": {
            "type": "string",
            "title": "Package Id"
          },
          "manifest_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Manifest Revision"
          },
          "primary_asset_id": {
            "type": "string",
            "title": "Primary Asset Id"
          },
          "model_asset_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 5,
            "minItems": 1,
            "title": "Model Asset Ids"
          },
          "incident_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Incident Version"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "fire_id",
          "episode_id",
          "package_id",
          "manifest_revision",
          "primary_asset_id",
          "model_asset_ids",
          "incident_version",
          "trace_id"
        ],
        "title": "AdminIncidentRepresentationAttachResponse"
      },
...
      "AdminIncidentSummary": {
        "properties": {
          "fire_id": {
            "type": "string",
            "title": "Fire Id"
          },
          "canonical_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Name"
          },
          "territory_code": {
            "type": "string",
            "title": "Territory Code"
          },
          "visibility": {
            "$ref": "#/components/schemas/PublicVisibility"
          },
          "current_episode_id": {
            "type": "string",
            "title": "Current Episode Id"
          },
          "status": {
            "$ref": "#/components/schemas/IncidentStatus"
          },
          "verification_state": {
            "$ref": "#/components/schemas/VerificationState"
          },
          "corroborating_source_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Corroborating Source Count"
          },
          "estimated_area_ha": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Area Ha"
          },
          "evacuation_established": {
            "type": "boolean",
            "title": "Evacuation Established"
          },
          "model_generation_eligible": {
            "type": "boolean",
            "title": "Model Generation Eligible"
          },
          "review_required": {
            "type": "boolean",
            "title": "Review Required"
          },
          "last_observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Observed At"
          },
          "pending_observation_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Pending Observation Count"
          },
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "fire_id",
          "territory_code",
          "visibility",
          "current_episode_id",
          "status",
          "verification_state",
          "corroborating_source_count",
          "evacuation_established",
          "model_generation_eligible",
          "review_required",
          "last_observed_at",
          "pending_observation_count",
          "version"
        ],
        "title": "AdminIncidentSummary"
      },
...
      "AdminOperationalMapResponse": {
        "properties": {
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "coordinate_system": {
            "type": "string",
            "const": "EPSG:4326",
            "title": "Coordinate System",
            "default": "EPSG:4326"
          },
          "summary": {
            "$ref": "#/components/schemas/AdminOperationalMapSummary"
          },
          "incidents": {
            "items": {
              "$ref": "#/components/schemas/AdminOperationalMapIncident"
            },
            "type": "array",
            "maxItems": 5000,
            "title": "Incidents"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "generated_at",
          "summary"
        ],
        "title": "AdminOperationalMapResponse"
      },
...
      "AdminWorkQueueIncident": {
        "properties": {
          "fire_id": {
            "type": "string",
            "title": "Fire Id"
          },
          "episode_id": {
            "type": "string",
            "title": "Episode Id"
          },
          "status": {
            "$ref": "#/components/schemas/IncidentStatus"
          },
          "verification_state": {
            "$ref": "#/components/schemas/VerificationState"
          },
          "last_observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Observed At"
          },
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "fire_id",
          "episode_id",
          "status",
          "verification_state",
          "last_observed_at",
          "version"
        ],
        "title": "AdminWorkQueueIncident"
      },
      "AdminWorkQueueObservation": {
        "properties": {
          "observation_id": {
            "type": "string",
            "title": "Observation Id"
          },
          "source_key": {
            "type": "string",
            "title": "Source Key"
          },
          "observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Observed At"
          },
          "longitude": {
            "type": "number",
            "title": "Longitude"
          },
          "latitude": {
            "type": "number",
            "title": "Latitude"
          },
          "horizontal_uncertainty_m": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Horizontal Uncertainty M"
          },
          "verification_state": {
            "$ref": "#/components/schemas/VerificationState"
          },
          "proposed_fire_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposed Fire Id"
          },
          "proposed_episode_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposed Episode Id"
          },
          "proposed_episode_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IncidentStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "match_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Score"
          },
          "review_reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Review Reasons"
          },
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "observation_id",
          "source_key",
          "observed_at",
          "longitude",
          "latitude",
          "horizontal_uncertainty_m",
          "verification_state",
          "version"
        ],
        "title": "AdminWorkQueueObservation"
      },
      "AdminWorkQueueResponse": {
        "properties": {
          "observations": {
            "items": {
              "$ref": "#/components/schemas/AdminWorkQueueObservation"
            },
            "type": "array",
            "maxItems": 200,
            "title": "Observations"
          },
          "reports": {
            "items": {
              "$ref": "#/components/schemas/PublicIncidentReport"
            },
            "type": "array",
            "maxItems": 200,
            "title": "Reports"
          },
          "incidents": {
            "items": {
              "$ref": "#/components/schemas/AdminWorkQueueIncident"
            },
            "type": "array",
            "maxItems": 200,
            "title": "Incidents"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AdminWorkQueueResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PublicIncidentReport": {
        "properties": {
          "report_id": {
            "type": "string",
            "title": "Report Id"
          },
          "fire_id": {
            "type": "string",
            "title": "Fire Id"
          },
          "category": {
            "$ref": "#/components/schemas/PublicReportCategory"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "state": {
            "$ref": "#/components/schemas/PublicReportState"
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Submitted At"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "closure_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closure Reason"
          },
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "report_id",
          "fire_id",
          "category",
          "message",
          "state",
          "submitted_at",
          "version"
        ],
        "title": "PublicIncidentReport"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}